+++ src/App.php
@@ -42,9 +42,9 @@ use Symfony\Component\Console\Output\OutputInterface;
 
 use function Busarm\PhpMini\Helpers\is_cli;
 
-// TODO:
-// Event Manager Interface - Handle sync and async dispatch
-// Queue Manager Interface - Handle sync and async jobs
+// TODO Event Manager Interface - Handle sync and async dispatch
+// TODO Queue Manager Interface - Handle sync and async jobs
+// TODO PSR Cache Interface
 /**
  * Application Factory
  *