+++ src/Data/PDO/Relations/OneToMany.php
@@ -3,7 +3,11 @@
 namespace Busarm\PhpMini\Data\PDO\Relations;
 
 use Busarm\PhpMini\Data\PDO\Model;
+use Busarm\PhpMini\Data\PDO\Reference;
 use Busarm\PhpMini\Data\PDO\Relation;
+use Busarm\PhpMini\Enums\DataType;
+
+use function Busarm\PhpMini\Helpers\is_list;
 
 /**
  * PHP Mini Framework
@@ -16,32 +20,27 @@ class OneToMany extends Relation
     /**
      * @param string $name Relation attribute name in Current Model
      * @param Model $model Current Model 
-     * @param Model $toModel Related Model
-     * @param array $references e.g. `['modelKey1' => 'toModelKey1', 'modelKey2' => 'toModelKey2']`
+     * @param Reference $reference Relation Reference
      * @return void
      */
     public function __construct(
         private string $name,
         private Model $model,
-        private Model $toModel,
-        private array $references
+        private Reference $reference
     ) {
-        parent::__construct($name, $toModel);
+        parent::__construct($name, DataType::ARRAY);
     }
 
     /**
      * Get relation data
      * 
-     * @param array $conditions
-     * @param array $params
-     * @param array $columns
      * @return Model[]
      */
-    public function get(array $conditions = [], array $params = [], array $columns = ['*']): array
+    public function get(): array
     {
         $referenceConditions = [];
         $referenceParams = [];
-        foreach ($this->references as $modelRef => $toModelRef) {
+        foreach ($this->getReferences() as $modelRef => $toModelRef) {
             if (isset($this->model->{$modelRef})) {
                 $referenceConditions[] = "$toModelRef = :$toModelRef";
                 $referenceParams[":$toModelRef"] = $this->model->{$modelRef};
@@ -49,10 +48,10 @@ class OneToMany extends Relation
         }
 
         if (count($referenceConditions) && count($referenceConditions)) {
-            return $this->toModel->setAutoLoadRelations(false)->all(
-                array_merge($referenceConditions, $conditions),
-                array_merge($referenceParams, $params),
-                $columns
+            return $this->getReferenceModel()->clone()->setAutoLoadRelations(false)->all(
+                array_merge($referenceConditions, $this->conditions),
+                array_merge($referenceParams, $this->params),
+                $this->columns
             );
         }
         return [];
@@ -62,17 +61,15 @@ class OneToMany extends Relation
      * Load relation data for list of items
      * 
      * @param Model[] $items
-     * @param array $conditions
-     * @param array $conditions
-     * @param array $params
-     * @param array $columns
      * @return Model[] $items with loaded relations
      */
-    public function load(array $items, array $conditions = [], array $params = [], array $columns = ['*']): array
+    public function load(array $items): array
     {
+        if (empty($items)) return [];
+
         $referenceConditions = [];
         $referenceParams = [];
-        foreach ($this->references as $modelRef => $toModelRef) {
+        foreach ($this->getReferences() as $modelRef => $toModelRef) {
             $refs = array_map(fn ($item) => $item->{$modelRef}, $items);
             if (!empty($refs)) {
                 $referenceConditions[] = sprintf("$toModelRef IN (%s)", implode(',', array_fill(0, count($refs), '?')));
@@ -83,22 +80,22 @@ class OneToMany extends Relation
         if (count($referenceConditions) && count($referenceConditions)) {
 
             // Get relation results for all items
-            $results = $this->toModel->setAutoLoadRelations(false)->all(
-                array_merge($referenceConditions, $conditions),
-                array_merge($referenceParams, $params),
-                $columns
+            $results = $this->getReferenceModel()->clone()->setAutoLoadRelations(false)->setPerPage($this->limit * count($items))->all(
+                array_merge($referenceConditions, $this->conditions),
+                array_merge($referenceParams, $this->params),
+                $this->columns
             );
 
             // Group result for references
             $resultsMap = [];
             foreach ($results as $result) {
-                $key = implode('-', array_map(fn ($ref) => $result->{$ref}, array_values($this->references)));
+                $key = implode('-', array_map(fn ($ref) => $result->{$ref}, array_values($this->getReferences())));
                 $resultsMap[$key][] = $result;  // Multiple items (1:m)
             }
 
             // Map relation for each item
             foreach ($items as &$item) {
-                $key = implode('-', array_map(fn ($ref) => $item->{$ref}, array_keys($this->references)));
+                $key = implode('-', array_map(fn ($ref) => $item->{$ref}, array_keys($this->getReferences())));
                 $item->{$this->getName()} = $resultsMap[$key] ?? [];
                 $item->addLoadedRelation($this->getName());
             }
@@ -108,22 +105,63 @@ class OneToMany extends Relation
         return [];
     }
 
-    /**
-     * Get relation references
-     * @return array
-     */
-    public function getReferences(): array
-    {
-        return $this->references;
-    }
 
     /**
-     * Get relation reference model
-     * @return Model
+     * Save relation data
+     * 
+     * @param array|array<array> $data Singe array item or 2D array multiple items
+     * @return bool
      */
-    public function getReferenceModel(): Model
+    public function save(array $data): bool
     {
-        return $this->toModel;
+        if (empty($data)) return false;
+
+        // Is multiple values
+        if (is_list($data)) {
+            return $this->getReferenceModel()->transaction(function () use ($data) {
+                $done = true;
+                foreach ($data as $item) {
+                    if (is_array($item)) {
+                        if (!$this->save($item)) $done = false;
+                    }
+                }
+                return $done;
+            });
+        }
+
+        $referenceModel = $this->getReferenceModel()->clone();
+
+        // Load reference model keys in to $data if available
+        $reFieldNames = $referenceModel->getFieldNames();
+        foreach ($this->getReferences() as $modelRef => $toModelRef) {
+            if (!isset($data[$toModelRef]) && isset($this->getCurrentModel()->{$modelRef}) && in_array($toModelRef, $reFieldNames)) {
+                $data[$toModelRef] = $this->getCurrentModel()->{$modelRef};
+            }
+        }
+
+        // Save reference model
+        $referenceModel->load($data);
+        if ($referenceModel->save()) {
+
+            $modelData = [];
+
+            // Load current model keys
+            $fieldNames = $this->getCurrentModel()->getFieldNames();
+            foreach ($this->getReferences() as $modelRef => $toModelRef) {
+                if (isset($referenceModel->{$toModelRef}) && in_array($modelRef, $fieldNames)) {
+                    $modelData[$modelRef] = $referenceModel->{$toModelRef};
+                }
+            }
+            // Save current model keys
+            if (!empty($modelData) || isset($this->getCurrentModel()->{$this->getCurrentModel()->getKeyName()})) {
+                $this->getCurrentModel()->load($modelData);
+                $this->getCurrentModel()->save(false, false);
+            }
+
+            return true;
+        }
+
+        return false;
     }
 
     /**
@@ -133,5 +171,23 @@ class OneToMany extends Relation
     public function getCurrentModel(): Model
     {
         return $this->model;
+    }
+
+    /**
+     * Get relation reference model
+     * @return Model
+     */
+    public function getReferenceModel(): Model
+    {
+        return $this->reference->getModel();
+    }
+
+    /**
+     * Get relation references
+     * @return array
+     */
+    public function getReferences(): array
+    {
+        return $this->reference->getKeys();
     }
 }