+++ tests/TestApp/Models/ProductTagTestModel.php
@@ -3,6 +3,7 @@
 namespace Busarm\PhpMini\Test\TestApp\Models;
 
 use Busarm\PhpMini\Data\PDO\Model;
+use Busarm\PhpMini\Data\PDO\Reference;
 use Busarm\PhpMini\Data\PDO\Relations\OneToMany;
 use Busarm\PhpMini\Data\PDO\Relations\OneToOne;
 
@@ -32,8 +33,8 @@ class ProductTagTestModel extends Model
     public function getRelations(): array
     {
         return [
-            new OneToOne('product', $this, new ProductTestModel, ['productId' => 'id']),
-            new OneToOne('tag', $this, new TagTestModel, ['tagId' => 'id']),
+            new OneToOne('product', $this, new Reference(new ProductTestModel, ['productId' => 'id'])),
+            new OneToOne('tag', $this, new Reference(new TagTestModel, ['tagId' => 'id'])),
         ];
     }